Fix: Delay publish workflow produces error and doesn't continue
The issue is that someone scheduled an article to be published. Then the article was either edited and republished 10 times (or more) before the publishing time happened, or one of the old versions of the object was deleted. Due to this the workflow crashes when it comes to this object and other waiting articles will not be published.

When we publish an article the workflow will start. It will wait for a specific time until next event strikes. If we edit the object and publish again then the same will happen on the new version. The workflow-event will then be running for both versions of the same object. The problem is that we have a 10 version limit in the database, so when we edit the object the 11th time the first version will be deleted. If this happens while version 1 is still having the event running and waiting for a date, then we will get an error. 

This should have been fixed in the kernel of eZ, but it is not.

This FIX will end all workflows running for removed versions and problem is solved.